server starts
