+++ controllers/api-routes.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+//const models = require('../../models');
 
 // ===============================================================================
 // ROUTING
@@ -6,33 +7,6 @@ const mongoose = require('mongoose');
 
 module.exports = function(app) {
 	
-	// If deployed, use the deployed database. Otherwise use the local mongoHeadlines database
-const MONGODB_URI = process.env.MONGODB_URI || "mongodb://localhost/invoiceCentral";
-
-mongoose.connect(MONGODB_URI, { useNewUrlParser: true });
-
-const db = mongoose.connection;
-db.on('error', console.error.bind(console, 'connection error:'));
-db.once('open', function() {
-	console.log("Connected to MongoDB.");
-  // we're connected!
-});
-	
-const Schema = mongoose.Schema;
-
-const customerSchema = new Schema({
-    firstName:  String,
-	lastName:  String,
-	nickName:  String,
-	address1: String,
-	address2: String,
-	city: String,
-	state: String,
-	zip: String,
-	country: String
-});
-
-const Customer = mongoose.model("Customer", customerSchema);
 
 /*
 let cust = new Customer({
@@ -49,17 +23,21 @@ let cust = new Customer({
 cust.save();
 */
 
-app.get("/api/fetch", function(req, res) {
+app.get("/fetch", function(req, res) {
 	console.log("get fetch requested.");
 	console.log("get all customers");
-	Customer.find({}).then( function(customers) {res.send(customers).status(200);});
+	res.send("Get Customers Route Hit").status(200);
+	
+	//models.Customer.find({}).then( function(customers) {res.send(customers).status(200);});
 });
 
-app.delete("/api/customer/*", function(req, res) {
+/*
+app.delete("/customer/*", function(req, res) {
 	Customer.findById(req.params[0], function (err, customer) { if(customer) customer.remove();}).then(function (customer){res.send(customer).status(200)});
 });
 
-app.post("/api/customer", function(req, res) {
+
+app.post("/customer", function(req, res) {
 	console.log(req.body);
 	let customer = new Customer(req.body).save().then(function(customer){
 		res.send(customer).status(200);
@@ -68,7 +46,7 @@ app.post("/api/customer", function(req, res) {
 	});
 });
 
-app.put("/api/customer/*", function(req, res) {
+app.put("/customer/*", function(req, res) {
 	console.log("update customer route hit");
 	console.log(req.params); //this specifies which customer record we are adding information to
 	console.log(req.body); //use this to specify properties and values to add
@@ -78,5 +56,6 @@ app.put("/api/customer/*", function(req, res) {
 		console.log(numAffected);
 	}).then(function (customer){console.log(customer);res.send(customer).status(200)});
 });
+*/
 
 };
