+++ models/index.js
@@ -1,0 +1,4 @@
+const customerRoutes = require("./customer.js");
+
+
+module.exports = customerRoutes;
